Add regression test for the data analysis bug, reorganize the classes into files
