+++ spec/requests/admin/data_analyses_requests_spec.rb
@@ -1,6 +1,17 @@
 require "rails_helper"
 
 RSpec.describe "Admin/Data Analysis requests" do
+  describe "GET /admin/data_analyses/:top_countries" do
+    it "returns the full country names in labels" do
+      student = FactoryBot.create(:student)
+
+      sign_in(:admin)
+      get admin_data_analysis_path(:top_countries)
+
+      expect(JSON.parse(response.body)['data']['attributes']['labels']).to eq(["United States"])
+    end
+  end
+
   describe "GET /admin/data_analyses/:returning_students" do
     it "counts returning students with multiple seasons" do
       student = FactoryBot.create(:student, :past)
